move schedule helper in <Resource/> to schedule.ts utils
