+++ src/components/Resource.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import moment from 'moment';
-import { TDay, TResource, TSchedule, TSchedulePeriod } from '../types';
+import { TResource, TSchedule } from '../types';
 import useResource from './useResource';
 import { getSearchParamVal } from '../utils/searchParams';
+import { orderWeeklySchedule, orderMonthlySchedule } from '../utils/schedule';
 import { SEARCH_PARAM_RESOURCE, FIREBASE_RESOURCE_BRANCH } from '../constants';
 import LoadingSpinner from './LoadingSpinner';
 import { Container } from '../App.styles';
@@ -19,57 +20,6 @@ const renderContactInfo = (resource: TResource) => {
       <p>{phone}</p>
     </>
   );
-};
-
-const days: TDay[] = [
-  'Sunday',
-  'Monday',
-  'Tuesday',
-  'Wednesday',
-  'Thursday',
-  'Friday',
-  'Saturday'
-];
-
-const periods: TSchedulePeriod[] = ['First', 'Second', 'Third', 'Fourth'];
-
-const orderWeeklySchedule = (schedule: TSchedule[]) => {
-  return schedule.sort(
-    (
-      { day: day1, tostring: fromstring1 },
-      { day: day2, fromstring: fromstring2 }
-    ) => {
-      if (day1 === day2) {
-        const startTime1 = parseInt(moment(fromstring1, 'h:mm A').format('H'));
-        const startTime2 = parseInt(moment(fromstring2, 'h:mm A').format('H'));
-        return startTime1 - startTime2;
-      }
-      return days.indexOf(day1) - days.indexOf(day2);
-    }
-  );
-};
-
-const orderMonthlySchedule = (schedule: TSchedule[]) => {
-  return schedule.sort((schedule1, schedule2) => {
-    const { period: period1, day: day1, fromstring: fromstring1 } = schedule1;
-    const { period: period2, day: day2, fromstring: fromstring2 } = schedule2;
-    if (period1 && period2) {
-      if (period1 === period2) {
-        if (day1 === day2) {
-          const startTime1 = parseInt(
-            moment(fromstring1, 'h:mm A').format('H')
-          );
-          const startTime2 = parseInt(
-            moment(fromstring2, 'h:mm A').format('H')
-          );
-          return startTime1 - startTime2;
-        }
-        return days.indexOf(day1) - days.indexOf(day2);
-      }
-      return periods.indexOf(period1) - periods.indexOf(period2);
-    }
-    return 0;
-  });
 };
 
 const generateSchedule = (schedule: TSchedule[]) => {
